backend/requirements.txt
django
django-rest-framework
psycopg2-binary
django-environ==0.11.1
channels==3.0.4 
channels-redis==3.3.1
msgpack-python==0.5.6
paho-mqtt==1.6.1
six==1.16.0
sqlparse==0.4.4
uvicorn[standard]
gunicorn==19.1.1
faker==19.1.0
celery==5.3.1
influxdb-client
redis
celery[redis]
django-filter
django-rest-swagger
drf-yasg
django-cors-headers
python-dotenv
factory-boy
Faker
ortools
llama-index
crewai
googlemaps